Database added
Added database logic to this bot. Will be doing testing after this
commit, but this is the initial implementation. It should now record all
trades and their statuses, whilst updating (insert, update, delete) the
records as the bot runs. This way, the bot can be stopped and restarted
and continue where it left off vs the issues I was having with in-memory
storage. SQL Server required.